# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  handlers.ChangePasswordRequest:
    properties:
      confirm_password:
        type: string
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
...
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  handlers.ResetPasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - confirm_password
    - new_password
    - token
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@forgecrud.com
    name: API Support
    url: http://www.forgecrud.com/support
  description: Central API documentation - For all ForgeCRUD microservices
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForgeCRUD API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change user's password after verifying current password
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request format or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated or incorrect password
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update password
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth-password
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiates password reset process by sending a reset link to the
        user's email
      parameters:
      - description: Email for password reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request format
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too many password reset attempts
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to process request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot password
      tags:
      - auth-password
...
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user's password using a valid reset token
      parameters:
      - description: Password reset data with token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request format or token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - auth-password
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...